Further refactoring;
- Remove instance variables which absent-mindedly crept in.
- Reduce duplication further by extracting resource retrieval out to module.
- Move ThorFixes class into CLI::Helpers for better organisation.
