+++ test/lib/troo/cli/cli_helpers_test.rb
@@ -3,6 +3,9 @@ require "thor"
 
 module Troo
   module CLI
+    module Helpers
+    end
+
     describe ThorFixes do
       let(:described_class) { ThorFixes }
